Server: Struktureller Umbau auf ES6 Modules
