+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "setup": "node setup.mjs",
-    "postinstall": "(cd server && npm install) && (node setup.mjs)"
+    "postinstall": "(cd server && npm install) && (node setup.mjs)",
+    "start": "cd server && npm start"
   },
   "repository": {
     "type": "git",
